requirements.txt
#devops
black==22.3.0
click==8.1.3 
pytest==7.1.3
pytest-cov==4.0.0
requests==2.26.0
fire==0.5.0
#rust based linter
ruff==0.0.284
# Azure Databricks SQL Query
python-dotenv=1.0.0
databricks-sql-connector=2.9.3